decoupled lap delta trace config class
