+++ Race_Element.HUD.ACC/Overlays/Driving/OverlayLapDeltaTrace/LapDeltaGraph.cs
@@ -4,9 +4,9 @@ using System.Collections.Generic;
 using System.Drawing;
 using System.Drawing.Drawing2D;
 using System.Linq;
-using static RaceElement.HUD.ACC.Overlays.Driving.OverlayLapDeltaGraph.LapDeltaTraceOverlay;
+using static RaceElement.HUD.ACC.Overlays.Driving.OverlayLapDeltaTrace.LapDeltaTraceOverlay;
 
-namespace RaceElement.HUD.ACC.Overlays.Driving.OverlayLapDeltaGraph
+namespace RaceElement.HUD.ACC.Overlays.Driving.OverlayLapDeltaTrace
 {
     internal class LapDeltaGraph : IDisposable
     {
